# openapi.yaml
openapi: 3.0.0
servers:
  - url: https://api.kalastud.io
    description: Production Server
  - url: https://perfect-week-test.herokuapp.com
    description: Test Server
info:
  title: PerfectWeek API
  description: The PerfectWeek web API
  version: 0.1.0
  contact:
    email: perfectweek_2020@labeip.epitech.eu
  license:
    name: MIT
    url: 'https://github.com/PerfectWeek/web-api/blob/dev/LICENSE'

####################
# Tags
#
...
paths:
...
  /auth/validate-email/{uuid}:
    get:
      tags:
        - Auth
      summary: Validate an email
      description: Finalize User creation
      operationId: confirmUserEmail
      parameters:
        - in: path
          name: uuid
          description: Unique identifier generated for email validation
          required: true
          schema:
            type: string
      responses:
        '201':
          description: User successfuly validated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: User has been successfully created
                  user:
                    $ref: '#/components/schemas/UserInfo'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /users:
    post:
      tags:
        - User
      summary: Create a new User
      description: The User will be created as "pending" until its email is verified
      operationId: createUser
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RegistrationCredentials'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: A link has been sent to your email address, please click on it in order to confirm your email
                  user:
                    $ref: '#/components/schemas/UserInfo'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid fields in User
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Pseudo or email already exists
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /users/{pseudo}:
    get:
      tags:
        - User
      summary: Get information about a specific User
      operationId: getUserInfo
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: pseudo
          required: true
          description: The User pseudo
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
                  user:
                    $ref: '#/components/schemas/UserInfo'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: User not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

    put:
      tags:
        - User
      summary: Edit User information
      operationId: editUserInfo
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: pseudo
          required: true
          description: The User pseudo
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserInfo'
      responses:
        '200':
          description: User successfully updated
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
                  user:
                    $ref: '#/components/schemas/UserInfo'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid fields in User
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '403':
          description: Action not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: User not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

    delete:
      tags:
        - User
      summary: Permanently delete a User
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: pseudo
          required: true
          description: The User pseudo
          schema:
            type: string
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: User deleted
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '403':
          description: Action not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /users/{pseudo}/upload-image:
    post:
      tags:
        - User
      summary: Upload image as new Avatar for own account
      operationId: uploadUserImage
      parameters:
      - in: path
        name: pseudo
        required: true
        description: The User pseudo
        schema:
          type: string
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
            encoding:
              image:
                contentType: image/png, image/jpeg
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
        '413':
          description: Entity Too Large
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Image should not exceed 2MB
        '404':
          description: User Not Found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: User not Found
        '400':
          description: File not Found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: File not Found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /users/{pseudo}/image:
    get:
      tags:
        - User
      summary: Get user profile image
      operationId: getUserImage
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: pseudo
          required: true
          description: The User pseudo
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
                  image:
                    type: string
                    example: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAfQAAAH0CAIAAABEtEjdAAABG2l...'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: User not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /users/{pseudo}/groups:
    get:
      tags:
        - User
      summary: Get all Groups a User belongs to
      operationId: getUserGroups
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: pseudo
          required: true
          description: The User pseudo
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/GroupRecap'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '403':
          description: Action not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: User not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /users/{pseudo}/calendars:
    get:
      tags:
        - User
      summary: Get all Calendars a User owns
      operationId: getUserCalendars
      description: >
        Retrieve all calendars of a User with its role in each.


        Role can be one of the following:

        - `admin` # Full access

        - `actor` # Can create new events

        - `spectator` # Can only view events

        - `outsider` # Can only view events it is invited to

      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: pseudo
          required: true
          description: The User pseudo
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
                  calendars:
                    type: array
                    items:
                      $ref: '#/components/schemas/CalendarRecap'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '403':
          description: Action not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: User not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

  #
  # Group
  #
...
  /calendars:
    post:
      tags:
        - Calendar
      summary: Create a new Calendar
      operationId: createCalendar
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CalendarCredentials'
      responses:
        '201':
          description: Calendar created
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Calendar created
                  calendar:
                    $ref: '#/components/schemas/CalendarInfo'
        '400':
          description: Invalid Calendar
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Invalid fields in Calendar
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /search/users:
    get:
      tags:
        - Search
      summary: Search users by pseudo
      operationId: searchUser
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: q
          required: true
          description: The pseudo to look for
          schema:
            type: string
        - in: query
          name: page_size
          required: false
          description: The max number of users to be fetched by page (must be between 1 and 20)
          schema:
            type: number
            default: 10
        - in: query
          name: page_number
          required: false
          description: The page from which to start fetching
          schema:
            type: number
            default: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: OK
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicUserInfo'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationFailed'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'


####################
# Component
#
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  #
  # Schemas
  #
  schemas:
...
    RegistrationCredentials:
      type: object
      properties:
        pseudo:
          type: string
          format: '[a-zA-Z][a-zA-Z0-9]*'
          example: amazing_user
        email:
          type: string
          example: amazing.user@gmail.com
        password:
          type: string
          example: AmazingPassword42
...
    UserInfo:
      type: object
      properties:
        pseudo:
          type: string
          example: amazing_user
        email:
          type: string
          example: amazing.user@gmail.com
        providers:
          type: object
          properties:
            facebook:
              type: boolean
              example: true
            google:
              type: boolean
              example: false
    PublicUserInfo:
      type: object
      properties:
        pseudo:
          type: string
          example: amazing_user
    GroupRecap:
      type: object
      properties:
        id:
          type: integer
          example: 42
        name:
          type: string
          example: amazing_group
        nb_members:
          type: number
          example: 3
        role:
          $ref: '#/components/schemas/GroupMemberRole'
...
    GroupMemberRole:
      type: string
      enum:
        - admin
        - spectator
        - actor
        - outsider
...
    CalendarRecap:
      type: object
      properties:
        calendar:
          type: object
          properties:
            id:
              type: integer
              example: 42
            name:
              type: string
              example: amazing_calendar
            role:
              $ref: '#/components/schemas/GroupMemberRole'
    CalendarCredentials:
      type: object
      properties:
        name:
          type: string
          example: amazing_calendar
    CalendarInfo:
      type: object
      properties:
        id:
          type: integer
          example: 42
        name:
          type: string
          example: amazing_calendar
        role:
          $ref: '#/components/schemas/GroupMemberRole'
...
    AuthenticationFailed:
      description: The authentication failed
      type: object
      properties:
        message:
          type: string
          example: You must be authenticated to perform this action
    # 403
    Unauthorized:
      description: The action is not allowed
      type: object
      properties:
        message:
          type: string
          example: Action not allowed
    # 404
    NotFound:
      description: Route or Resource not found
      type: object
      properties:
        message:
          type: string
          example: Route or Resource not found
    # 500
    InternalServerError:
      description: Internal Server Error
      type: object
      properties:
        message:
          type: string
          example: Internal Server Error